requirements.txt
discord.py==2.3.2
aiosqlite==0.19.0
python-dotenv==1.0.0
requests==2.31.0
pillow==10.1.0
aiohttp==3.9.1
urllib3==2.1.0
wavelink==2.6.4
youtube_dl==2021.12.17
asyncio==3.4.3 